requirements.txt
asgiref==3.6.0
Babel==2.11.0
Django==4.1.7
django-admin-rangefilter==0.9.0
django-dbml==0.6.0
django-extensions==3.2.1
django-filter==22.1
django-money==3.0.0
djangorestframework==3.12.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
py-moneyed==2.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
template==0.7.6
toml==0.10.2
typing_extensions==4.4.0
tzdata==2022.7
